Support caffe2 distribution training in Yarn.
For apps that requires gpu, the gpu allocation result will be saved in
the mpiExecDir.

Defeat:
If there are more than one containers on the same host for the same app,
the gpu allocation result file will be overwritten.
In order to solve this problem, we could:
1, Only one container is allowed on one host and merge all the small
containers;
2, Only one container is allowed on one host and discard the other
containers;
3, Find a way to make apps know that which gpus they have acquired;

Notice:
  Make sure that you use hadoop from
  https://github.com/Applepleple/hadoop.git with branch
  Applepleple/branch-2.7.2-Caffe2OnYarn
  or the build would fail.